requirements.txt
prettytable~=3.12.0
PyYAML~=6.0.2
